feat: refactor: Extract JWT utils to api-gateway
