+++ stakeholders-service/go.mod
@@ -5,7 +5,7 @@ go 1.24.5
 require (
 	api-gateway v0.0.0-00010101000000-000000000000
 	github.com/gin-gonic/gin v1.10.1
-	github.com/golang-jwt/jwt/v5 v5.2.3
+	github.com/golang-jwt/jwt/v5 v5.3.0
 	github.com/neo4j/neo4j-go-driver/v5 v5.28.2
 	go.mongodb.org/mongo-driver v1.17.4
 	golang.org/x/crypto v0.40.0
