+++ tours-service/handlers/keypoint_handler.go
@@ -1,13 +1,13 @@
 package handlers
 
 import (
+	utils "api-gateway/utils"
 	"fmt"
 	"net/http"
 	"path/filepath"
 	"time"
 	"tours-service/database"
 	"tours-service/models"
-	"tours-service/utils"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
@@ -16,8 +16,8 @@ import (
 func CreateKeyPoint(c *gin.Context) {
 	_, err := utils.VerifyJWT(c)
 	if err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-			return
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
 	}
 
 	var input struct {
