+++ tours-service/handlers/tour_handler.go
@@ -1,6 +1,7 @@
 package handlers
 
 import (
+	utils "api-gateway/utils"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -8,7 +9,6 @@ import (
 	"time"
 	"tours-service/database"
 	"tours-service/models"
-	"tours-service/utils"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/datatypes"
@@ -17,14 +17,14 @@ import (
 func CreateTour(c *gin.Context) {
 	claims, err := utils.VerifyJWT(c)
 	if err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-			return
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
 	}
 
 	userId, ok := claims["userId"].(string)
 	if !ok {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid userId in token"})
-			return
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid userId in token"})
+		return
 	}
 
 	name := c.PostForm("name")
@@ -41,8 +41,8 @@ func CreateTour(c *gin.Context) {
 
 	var tags []string
 	if err := json.Unmarshal([]byte(tagsStr), &tags); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid tags"})
-			return
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid tags"})
+		return
 	}
 
 	tour := models.Tour{
@@ -56,8 +56,8 @@ func CreateTour(c *gin.Context) {
 	}
 
 	if err := database.GORM_DB.Create(&tour).Error; err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save tour"})
-			return
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save tour"})
+		return
 	}
 
 	var keypointsInput []struct {
@@ -68,61 +68,61 @@ func CreateTour(c *gin.Context) {
 	}
 
 	if err := json.Unmarshal([]byte(c.PostForm("keypoints")), &keypointsInput); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid keypoints"})
-			return
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid keypoints"})
+		return
 	}
 
 	var keypoints []models.KeyPoint
 	for i, kp := range keypointsInput {
-			file, _ := c.FormFile(fmt.Sprintf("image%d", i))
+		file, _ := c.FormFile(fmt.Sprintf("image%d", i))
 
-			if kp.Name == "" || kp.Latitude == nil || kp.Longitude == nil || file == nil {
-				continue
-			}
+		if kp.Name == "" || kp.Latitude == nil || kp.Longitude == nil || file == nil {
+			continue
+		}
 
-			var imagePath string
-			filename := fmt.Sprintf("%d_%s", time.Now().UnixNano(), filepath.Base(file.Filename))
-			savePath := filepath.Join("static/uploads", filename)
-			if err := c.SaveUploadedFile(file, savePath); err != nil {
-					c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save keypoint image"})
-					return
-			}
-			imagePath = "/" + savePath
+		var imagePath string
+		filename := fmt.Sprintf("%d_%s", time.Now().UnixNano(), filepath.Base(file.Filename))
+		savePath := filepath.Join("static/uploads", filename)
+		if err := c.SaveUploadedFile(file, savePath); err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save keypoint image"})
+			return
+		}
+		imagePath = "/" + savePath
 
-			keypoints = append(keypoints, models.KeyPoint{
-					Name:        kp.Name,
-					Description: kp.Description,
-					Latitude:    *kp.Latitude,
-					Longitude:   *kp.Longitude,
-					ImagePath:   imagePath,
-					TourID:      tour.ID,
-			})
+		keypoints = append(keypoints, models.KeyPoint{
+			Name:        kp.Name,
+			Description: kp.Description,
+			Latitude:    *kp.Latitude,
+			Longitude:   *kp.Longitude,
+			ImagePath:   imagePath,
+			TourID:      tour.ID,
+		})
 	}
 
 	if len(keypoints) > 0 {
-			if err := database.GORM_DB.Create(&keypoints).Error; err != nil {
-					c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save keypoints"})
-					return
-			}
+		if err := database.GORM_DB.Create(&keypoints).Error; err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to save keypoints"})
+			return
+		}
 	}
 
 	c.JSON(http.StatusCreated, gin.H{
-			"tour":      tour,
-			"keypoints": keypoints,
+		"tour":      tour,
+		"keypoints": keypoints,
 	})
 }
 
 func GetAllTours(c *gin.Context) {
 	claims, err := utils.VerifyJWT(c)
 	if err != nil {
-			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-			return
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
 	}
 
 	userId, ok := claims["userId"].(string)
 	if !ok {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid userId in token"})
-			return
+		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid userId in token"})
+		return
 	}
 
 	var tours []models.Tour
